+++ yggdrasil/examples/model_function/src/model_function_modelB.jl
@@ -1,0 +1,6 @@
+function model_function(in_buf)
+  println("Model B:")
+  println(in_buf)
+  out_buf = in_buf
+  return out_buf
+end