+++ yggdrasil/examples/split_and_merge/src/split_and_merge_modelA.jl
@@ -1,0 +1,8 @@
+using Printf
+
+function modelA_function(in_val)
+  out_val1 = 2 * in_val
+  out_val2 = 3 * in_val
+  @printf("modelA_function(%s) = (%s, %s)\n", in_val, out_val1, out_val2)
+  return out_val1, out_val2
+end