+++ yggdrasil/examples/transformed_io/src/transformed_io_modelB.jl
@@ -1,0 +1,6 @@
+using Printf
+function modelB_function(in_val)
+  out_val = 3 * in_val
+  @printf("modelB_function(%s) = %s\n", in_val, out_val)
+  return in_val, out_val
+end