+++ yggdrasil/examples/transformed_io/src/transformed_io_modelC.jl
@@ -1,0 +1,6 @@
+using Printf
+function modelC_function(in_val)
+  out_val = 2 * in_val
+  @printf("modelC_function(%s) = %s\n", in_val, out_val)
+  return in_val, out_val
+end